fix: better coverage detection using anchors
Previously coverage was based on the start of AST nodes
in the source code. With the anchoring system, opcodes in
relevant source maps are searched, and the first one
within the source range of the AST node is selected
as an anchor.

If the anchor is executed once, then the coverage item
is considered to be covered. This works very well, but
false positives and edge cases still need to be found.